Update int from rc/2017-05-30

+++ app/services/idv/session.rb
@@ -1,16 +1,20 @@
 module Idv
   class Session
-    VALID_SESSION_ATTRIBUTES = [
-      :applicant,
-      :financials_confirmation,
-      :params,
-      :phone_confirmation,
-      :profile_confirmation,
-      :profile_id,
-      :recovery_code,
-      :resolution,
-      :step_attempts,
-      :vendor
+    VALID_SESSION_ATTRIBUTES = %i[
+      address_verification_mechanism
+      applicant
+      financials_confirmation
+      normalized_applicant_params
+      params
+      phone_confirmation
+      pii
+      profile_confirmation
+      profile_id
+      personal_key
+      resolution_successful
+      step_attempts
+      vendor
+      vendor_session_id
     ].freeze
 
     def initialize(user_session, current_user)
@@ -35,13 +39,14 @@ module Idv
     end
 
     def proofing_started?
-      resolution.present? && applicant.present? && resolution.success?
+      applicant.present? && resolution_successful
     end
 
-    def cache_applicant_profile_id(applicant)
-      profile = Idv::ProfileFromApplicant.create(applicant, current_user)
+    def cache_applicant_profile_id
+      profile = profile_maker.profile
+      self.pii = profile_maker.pii_attributes
       self.profile_id = profile.id
-      self.recovery_code = profile.recovery_code
+      self.personal_key = profile.personal_key
     end
 
     def cache_encrypted_pii(password)
@@ -49,9 +54,8 @@ module Idv
       cacher.save(password, profile)
     end
 
-    def applicant_from_params
-      app_vars = params.select { |key, _value| Proofer::Applicant.method_defined?(key) }
-      Proofer::Applicant.new(app_vars.merge(uuid: current_user.uuid))
+    def vendor_params
+      applicant_params_ascii.merge(uuid: current_user.uuid)
     end
 
     def profile
@@ -62,15 +66,31 @@ module Idv
       user_session.delete(:idv)
     end
 
+    def complete_session
+      complete_profile if phone_confirmation == true
+      create_usps_entry if address_verification_mechanism == 'usps'
+    end
+
     def complete_profile
       profile.verified_at = Time.zone.now
-      profile.vendor = vendor
       profile.activate
       move_pii_to_user_session
     end
 
+    def create_usps_entry
+      move_pii_to_user_session
+      if pii.is_a?(String)
+        self.pii = Pii::Attributes.new_from_json(user_session[:decrypted_pii])
+      end
+      UspsConfirmationMaker.new(pii: pii).perform
+    end
+
     def alive?
       session.present?
+    end
+
+    def address_mechanism_chosen?
+      phone_confirmation == true || address_verification_mechanism == 'usps'
     end
 
     private
@@ -82,11 +102,30 @@ module Idv
     end
 
     def move_pii_to_user_session
+      return if session[:decrypted_pii].blank?
       user_session[:decrypted_pii] = session.delete(:decrypted_pii)
     end
 
     def session
       user_session[:idv]
+    end
+
+    def applicant_params
+      params.select { |key, _value| Proofer::Applicant.method_defined?(key) }
+    end
+
+    def applicant_params_ascii
+      Hash[applicant_params.map { |key, value| [key, value.to_ascii] }]
+    end
+
+    def profile_maker
+      @_profile_maker ||= Idv::ProfileMaker.new(
+        applicant: Proofer::Applicant.new(applicant_params),
+        normalized_applicant: Proofer::Applicant.new(normalized_applicant_params),
+        phone_confirmed: phone_confirmation || false,
+        user: current_user,
+        vendor: vendor
+      )
     end
   end
 end
